finisce di implementare login e register, crea il metodo getItemForLoggedUser in MarketController

+++ src/main/java/org/generation/italy/reshare/model/repositories/abstractions/CityRepository.java
@@ -3,5 +3,5 @@ package org.generation.italy.reshare.model.repositories.abstractions;
 import org.generation.italy.reshare.model.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityRepository extends JpaRepository<City, Integer> {
+public interface CityRepository extends JpaRepository<City, Long> {
 }
