+++ src/main/java/org/generation/italy/reshare/model/repositories/abstractions/ReviewRepository.java
@@ -3,5 +3,5 @@ package org.generation.italy.reshare.model.repositories.abstractions;
 import org.generation.italy.reshare.model.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewRepository extends JpaRepository<Review, Integer> {
+public interface ReviewRepository extends JpaRepository<Review, Long> {
 }
