+++ src/main/java/org/generation/italy/reshare/model/services/abstractions/MarketService.java
@@ -7,13 +7,13 @@ import java.util.List;
 import java.util.Optional;
 
 public interface MarketService {
-List<Item> searchItemsByCategory(int categoryId) throws EntityNotFoundException;
+List<Item> searchItemsByCategory(long categoryId) throws EntityNotFoundException;
 //List<Item> searchItemsByCity(int cityId);
-List<Item> searchItemsByUser(int userId) throws EntityNotFoundException;
+List<Item> searchItemsByUser(long userId) throws EntityNotFoundException;
 //List<Item> searchItemsByName(String itemTypeName);
 //List<Item> searchOfferedItems();
 //Optional<AppUser> findUserById (int id);
 List<Item> searchItems(String condition, Boolean activetrade, Integer lastN);
 
-Item findItemById(int id) throws EntityNotFoundException;
+Item findItemById(long id) throws EntityNotFoundException;
 }
