+++ src/main/java/org/generation/italy/reshare/model/services/abstractions/WishlistService.java
@@ -7,14 +7,14 @@ import java.util.List;
 import java.util.Optional;
 
 public interface WishlistService {
-    boolean addItemTypeById(int userId, int itemTypeId);
-    boolean removeItemTypeById(int userId, int itemTypeId);
-    boolean clearWishlist(int userId);
-    AppUser getUserById(int id);
-    Optional<ItemType> getItemTypeById(int id);
+    boolean addItemTypeById(long userId, long itemTypeId);
+    boolean removeItemTypeById(long userId, long itemTypeId);
+    boolean clearWishlist(long userId);
+    AppUser getUserById(long id);
+    Optional<ItemType> getItemTypeById(long id);
     List<ItemType> getAllItemType();
-    List<ItemType> getAllItemTypeByCategory(int id);
-    List<ItemType> getAllItemTypeByCity(int id);
+    List<ItemType> getAllItemTypeByCategory(long id);
+    List<ItemType> getAllItemTypeByCity(long id);
 
 
 }
